split tests and 100 coverage
