+++ lab_6/plati_digiseller_parser/tests/conftest.py
@@ -7,3 +7,19 @@ SRC = ROOT / "src"
 for path in (ROOT, SRC):
     if str(path) not in sys.path:
         sys.path.append(str(path))
+
+
+class FakeResponse:
+    def __init__(self, text="", json_data=None, status_code=200):
+        self.text = text
+        self._json = json_data or {}
+        self.status_code = status_code
+        self.raise_called = False
+
+    def raise_for_status(self):
+        self.raise_called = True
+        if self.status_code >= 400:
+            raise Exception("error")
+
+    def json(self):
+        return self._json
