+++ lab_6/plati_digiseller_parser/tests/test_base_parser.py
@@ -1,0 +1,12 @@
+import pytest
+
+from src.parsers.base import Parser
+
+
+def test_base_parser_not_implemented():
+    class Dummy(Parser):
+        def fetch(self, config):
+            return super().fetch(config)
+
+    with pytest.raises(NotImplementedError):
+        Dummy().fetch(None)
